+++ packages/azure-kusto-data/src/clientDetails.ts
@@ -14,8 +14,8 @@ declare namespace NodeJS {
     }
 }
 
-// REPLACE_REGEX = re.compile(r"[\r\n\s{}|]+")
-const ReplaceRegex = /[\r\n\s{}|]+/g;
+// This regex allows all printable ascii, except spaces and chars we use in the format
+const ReplaceRegex = /[^\w.\-()]/g;
 const None = "[none]";
 
 export class ClientDetails {
@@ -33,7 +33,7 @@ export class ClientDetails {
 
     static defaultApplication(): string {
         if (isNode) {
-            return process.env?.npm_package_name || process.title || None;
+            return process?.env?.npm_package_name || process?.argv[1] || None;
         } else {
             return window?.location?.href || None;
         }
@@ -53,7 +53,7 @@ export class ClientDetails {
                 }
             }
 
-            return username || (process.env.USERDOMAIN ? `${process.env.USERDOMAIN}\\${process.env.USERNAME}` : process.env.USERNAME) || None;
+            return username || (process.env?.USERDOMAIN ? `${process.env?.USERDOMAIN}\\${process.env?.USERNAME}` : process?.env?.USERNAME) || None;
         } else {
             return None;
         }
@@ -66,8 +66,9 @@ export class ClientDetails {
         ]);
     }
 
-    static escapeHeader(header: string): string {
-        return `{${header.replace(ReplaceRegex, "_")}}`;
+    static escapeHeader(header: string, wrapInBrackets: boolean = true): string {
+        const clean = header.substring(0, 100).replace(ReplaceRegex, "_");
+        return wrapInBrackets ? `{${clean}}` : clean;
     }
 
     static formatHeader(args: [string, string][]): string {
@@ -86,7 +87,7 @@ export class ClientDetails {
         override_user: string | null = null,
         additional_fields: [string, string][] | null = null
     ): ClientDetails {
-        const params: [string, string][] = [["Kusto." + name, version]];
+        const params: [string, string][] = [["Kusto." + this.escapeHeader(name, false), version]];
 
         app_name = app_name || this.defaultApplication();
         app_version = app_version || None;