+++ packages/azure-kusto-ingest/src/ingestClientBase.ts
@@ -3,7 +3,7 @@
 
 import { Client as KustoClient, KustoConnectionStringBuilder } from "azure-kusto-data";
 
-import ResourceManager, { createStatusTableClient } from "./resourceManager";
+import ResourceManager from "./resourceManager";
 
 import IngestionBlobInfo from "./ingestionBlobInfo";
 import { ContainerClient } from "@azure/storage-blob";
@@ -27,14 +27,29 @@ import { BlobDescriptor, StreamDescriptor } from "./descriptors";
 
 export abstract class KustoIngestClientBase extends AbstractKustoClient {
     resourceManager: ResourceManager;
-
+    applicationForTracing: string | null;
+    clientVersionForTracing: string | null;
     static readonly MaxNumberOfRetryAttempts = 3;
 
-    constructor(kcsb: string | KustoConnectionStringBuilder, defaultProps?: IngestionPropertiesInput, isBrowser?: boolean) {
+    constructor(
+        kcsb: string | KustoConnectionStringBuilder,
+        defaultProps?: IngestionPropertiesInput,
+        autoCorrectEndpoint: boolean = true,
+        isBrowser?: boolean
+    ) {
         super(defaultProps);
+        if (typeof kcsb === "string") {
+            kcsb = new KustoConnectionStringBuilder(kcsb);
+        }
+        if (autoCorrectEndpoint) {
+            kcsb.dataSource = this.getIngestionEndpoint(kcsb.dataSource);
+        }
         const kustoClient = new KustoClient(kcsb);
         this.resourceManager = new ResourceManager(kustoClient, isBrowser);
         this.defaultDatabase = kustoClient.defaultDatabase;
+        const clientDetails = kcsb.clientDetails();
+        this.applicationForTracing = clientDetails.applicationNameForTracing;
+        this.clientVersionForTracing = clientDetails.versionForTracing;
     }
 
     async ingestFromBlob(
@@ -43,13 +58,12 @@ export abstract class KustoIngestClientBase extends AbstractKustoClient {
         maxRetries: number = KustoIngestClientBase.MaxNumberOfRetryAttempts
     ): Promise<IngestionResult> {
         this.ensureOpen();
-
         const props = this._getMergedProps(ingestionProperties);
 
         const descriptor = blob instanceof BlobDescriptor ? blob : new BlobDescriptor(blob);
 
         const authorizationContext = await this.resourceManager.getAuthorizationContext();
-        const ingestionBlobInfo = new IngestionBlobInfo(descriptor, props, authorizationContext);
+        const ingestionBlobInfo = new IngestionBlobInfo(descriptor, props, authorizationContext, this.applicationForTracing, this.clientVersionForTracing);
 
         const reportToTable = props.reportLevel !== ReportLevel.DoNotReport && props.reportMethod !== ReportMethod.Queue;